refactor: 비로그인 유저 알럿 처리
