+++ src/components/common/alert/Alert.tsx
@@ -13,7 +13,7 @@ const Alert: FC<AlertProps> = ({ title, message, onConfirm, children }) => (
     <Dialog.Trigger asChild>{children}</Dialog.Trigger>
 
     <Dialog.Portal>
-      <Dialog.Overlay className="fixed inset-0 z-10 bg-black opacity-70" />
+      <Dialog.Overlay className="fixed inset-0 z-10 h-[100%] bg-black opacity-70" />
       <Dialog.Content className="fixed left-[50%] top-[50%] z-10 flex h-[192px] w-[346px] translate-x-[-50%] translate-y-[-50%] flex-col items-center justify-center rounded-[16px] bg-white shadow-lg focus:outline-none">
         <Dialog.Description className="mt-[10px] text-center text-[15px] leading-normal">
           <p className="mb-[16px] font-bold text-[#062139]">{title}</p>
