+++ src/components/common/like/Like.tsx
@@ -3,9 +3,14 @@ import { HeartIcon } from '../icons/Icons';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { deleteLikedTours, postLikedTours } from '@api/tours';
 
+import Alert from '../alert/Alert';
+import { useState } from 'react';
+
 const Like = ({ liked, id }: LikeProps) => {
   const navigate = useNavigate();
   const queryClient = useQueryClient();
+
+  const [isLoggedIn, setIsLoggedIn] = useState<boolean>(false);
 
   const { mutate: likeMutate } = useMutation({
     mutationFn: (id: number) => postLikedTours({ id }),
@@ -30,27 +35,59 @@ const Like = ({ liked, id }: LikeProps) => {
   const onClickLikeButton = (e: React.MouseEvent<HTMLElement>) => {
     e.stopPropagation();
     const token = localStorage.getItem('accessToken');
-    if (!token) {
-      // 비로그인 알람창 처리 필요
-      navigate('/signin');
-    } else {
+    if (token) {
+      setIsLoggedIn(true);
       if (liked === false) {
         likeMutate(id);
       } else {
         unlikeMutate(id);
       }
+    } else {
+      setIsLoggedIn(false);
     }
   };
 
+  const handleConfirm = () => {
+    navigate('/login');
+  };
+
   return (
-    <div
-      onClick={onClickLikeButton}
-      className="top-75 h-[24px] w-[24px] cursor-pointer">
-      <HeartIcon
-        fill={liked ? '#FF2167' : '#D7D7D7'}
-        color={liked ? '#ff2167' : '#ffffff'}
-      />
-    </div>
+    <>
+      {isLoggedIn ? (
+        <>
+          <div
+            onClick={onClickLikeButton}
+            className="top-75 h-[24px] w-[24px] cursor-pointer">
+            <HeartIcon
+              fill={liked ? '#FF2167' : '#D7D7D7'}
+              color={liked ? '#ff2167' : '#ffffff'}
+            />
+          </div>
+        </>
+      ) : (
+        <>
+          <Alert
+            title={'로그인'}
+            message={
+              <>
+                관심 여행지 등록 및 삭제를 위해 로그인이 필요합니다.
+                <br />
+                로그인 하시겠습니까?
+              </>
+            }
+            onConfirm={handleConfirm}>
+            <div
+              onClick={onClickLikeButton}
+              className="top-75 h-[24px] w-[24px] cursor-pointer">
+              <HeartIcon
+                fill={liked ? '#FF2167' : '#D7D7D7'}
+                color={liked ? '#ff2167' : '#ffffff'}
+              />
+            </div>
+          </Alert>
+        </>
+      )}
+    </>
   );
 };
