+++ src/components/common/nav/Nav.tsx
@@ -1,11 +1,28 @@
 import { useLocation, useNavigate } from 'react-router-dom';
 import { CalendarIcon, HeartIcon, HomeIcon, UserIcon } from '../icons/Icons';
+import { useEffect, useState } from 'react';
+import Alert from '../alert/Alert';
 
 const Nav = () => {
   const navigate = useNavigate();
   const location = useLocation();
 
   const isActive = (path: string) => location.pathname === path;
+
+  const [isLoggedIn, setIsLoggedIn] = useState<boolean>(false);
+
+  useEffect(() => {
+    const token = window.localStorage.getItem('accessToken');
+    if (token) {
+      setIsLoggedIn(true);
+    } else {
+      setIsLoggedIn(false);
+    }
+  }, []);
+
+  const handleConfirm = () => {
+    navigate('/login');
+  };
 
   return (
     <nav className="sticky bottom-0 z-50 mt-auto flex h-16 items-center justify-center bg-white">
@@ -24,14 +41,40 @@ const Nav = () => {
           <CalendarIcon />
           <p className="caption1 mt-[5px] text-center text-xs/[11px]">일정</p>
         </div>
-        <div
-          onClick={() => navigate('/wishList')}
-          className="cursor-pointer flex-col items-center justify-center px-2">
-          <div className="flex justify-center">
-            <HeartIcon fill={isActive('/wishList') ? 'currentColor' : 'none'} />
+
+        {isLoggedIn ? (
+          <div
+            onClick={() => navigate('/wishList')}
+            className="cursor-pointer flex-col items-center justify-center px-2">
+            <div className="flex justify-center">
+              <HeartIcon
+                fill={isActive('/wishList') ? 'currentColor' : 'none'}
+              />
+            </div>
+            <p className="caption1 mt-[4px] text-center text-xs/[11px]">찜</p>
           </div>
-          <p className="caption1 mt-[4px] text-center text-xs/[11px]">찜</p>
-        </div>
+        ) : (
+          <Alert
+            title={'로그인'}
+            message={
+              <>
+                관심 여행지 조회를 위해 로그인이 필요합니다.
+                <br />
+                로그인 하시겠습니까?
+              </>
+            }
+            onConfirm={handleConfirm}>
+            <div className="cursor-pointer flex-col items-center justify-center px-2">
+              <div className="flex justify-center">
+                <HeartIcon
+                  fill={isActive('/wishList') ? 'currentColor' : 'none'}
+                />
+              </div>
+              <p className="caption1 mt-[4px] text-center text-xs/[11px]">찜</p>
+            </div>
+          </Alert>
+        )}
+
         <div
           onClick={() => navigate('/signin')}
           className="cursor-pointer flex-col items-center justify-center px-1 pt-[4.5px]">
